Past orders and checkout functionality completed

+++ store/views.py
@@ -1,4 +1,5 @@
 # Create your views here.
+from decimal import *
 from django.http import HttpResponse, HttpResponseRedirect
 from django.shortcuts import render
 from django.contrib.auth.models import User, Group
@@ -70,7 +71,7 @@ def login(request):
 
 def logout(request):
 	auth_logout(request)
-	return HttpResponse("You logged out brah!")
+	return redirect('store.index')
 
 @login_required(login_url='/store/login')
 def cart(request):
@@ -84,7 +85,6 @@ def cart(request):
 
 # QA:
 # 1. handle adding non-existant products
-# quantity of 0 will indicate we want to remove the item from the cart
 @login_required(login_url='/store/login')
 def cartadd(request, product_id):
 	quantity = int(request.POST.get('quantity', 1))
@@ -97,11 +97,10 @@ def cartadd(request, product_id):
 	if (created):
 		cart_item.quantity = quantity
 		cart_item.save()
-		return redirect('store.cart')
 	else:
 		cart_item.quantity += quantity
 		cart_item.save()
-		return HttpResponse("Yo, I added " + str(quantity) + " more " + product.name + "(s) to your cart to make " + str(cart_item.quantity))
+	return redirect('store.cart')	
 
 @login_required(login_url='/store/login')
 def cartupdate(request, product_id):
@@ -112,11 +111,10 @@ def cartupdate(request, product_id):
 	if quantity <= 0:
 		# Remove from cart
 		CartItem.objects.filter(shopping_cart=shopping_cart, product=product).delete()
-		return redirect('store.cart')
 	else:
 		cart_item.quantity = quantity
 		cart_item.save() 
-		return redirect('store.cart')
+	return redirect('store.cart')
 
 @login_required(login_url='/store/login')
 def cartremove(request, product_id):
@@ -126,29 +124,58 @@ def cartremove(request, product_id):
 	return redirect('store.cart')
 
 
-# QA:
-# 1. verify we have at least one product to checkout
+#QA:
+# 1. verify there are items in the cart if we are POSTing
 @login_required(login_url='/store/login')
 @transaction.commit_on_success
 def checkout(request):
-	DEBUG_empty_cart = True 
-	shopping_cart = ShoppingCart.objects.get(user=request.user, merchant=request.Merchant)
-	cart_items = CartItem.objects.filter(shopping_cart=shopping_cart)
+	if request.method == 'POST':
+		order = Order(order_date=datetime.datetime.now(), user=request.user, merchant=request.Merchant)
+		credit_card_form = CreditCardForm(request.POST, error_class=DivErrorList, instance=order)
+		shipping_info_form = ShippingInfoForm(request.POST, error_class=DivErrorList, instance=order)
 
-	# Create a new order
-	order = Order(order_date=datetime.datetime.now(), user=request.user, merchant=request.Merchant)
-	order.save()
+		if credit_card_form.is_valid() and shipping_info_form.is_valid():
+			
+			# Place the actual order
+			order.save()
+			DEBUG_empty_cart = False 
+			shopping_cart = ShoppingCart.objects.get(user=request.user, merchant=request.Merchant)
+			cart_items = CartItem.objects.filter(shopping_cart=shopping_cart)
 
-	order_items = []
-	for product in shopping_cart.products.all():
-		product_quantity = cart_items.get(product=product).quantity
-		order_items.append(OrderItem(order=order, product=product, quantity=product_quantity))
+			# Add all the shopping cart items to the order
+			order_items = []
+			order_total = Decimal(0.0)
+			for product in shopping_cart.products.all():
+				product_quantity = cart_items.get(product=product).quantity
+				product_price = cart_items.get(product=product).product.price
+				order_total += product_quantity * product_price
+				order_items.append(OrderItem(order=order, product=product, quantity=product_quantity))
 
-	OrderItem.objects.bulk_create(order_items)
+			order.total = order_total
+			order.save()
+			OrderItem.objects.bulk_create(order_items)
 
-	# Empty the shopping cart
-	if (DEBUG_empty_cart):
-		cart_items.delete()
+			# Empty the shopping cart
+			if (DEBUG_empty_cart):
+				cart_items.delete()
 
-	return HttpResponse("Yo, you successfully placed an order!")
+			return render(request, request.Merchant.subdomain + '/orderplaced.html')
+
+		else:
+			return render(request, request.Merchant.subdomain + '/checkout.html', {'credit_card_form':credit_card_form, 'shipping_info_form':shipping_info_form})
+
+	else:
+		credit_card_form = CreditCardForm()
+		shipping_info_form = ShippingInfoForm()
+		return render(request, request.Merchant.subdomain + '/checkout.html', {'credit_card_form':credit_card_form, 'shipping_info_form':shipping_info_form})
+
+# QA:
+# 1. verify we have at least one order to view
+@login_required(login_url='/store/login')
+def vieworders(request):
+	orders = Order.objects.filter(user=request.user, merchant=request.Merchant)
+	for order in orders:
+		order_items = OrderItem.objects.filter(order=order)
+		order.items = order_items
+	return render(request, request.Merchant.subdomain + '/vieworders.html', {'orders':orders})
